+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- voice notifications upon spawn
+
+### Changed
+- react-progress-timer dep (v0.2.0 --> v0.3.0) for onFinish
+- ToggleTipButton extracted from VoiceCommandButton
 
 ## [0.2.1] - 2022-06-05
 ### Changed